Rudimentary corruption implemented
When a player's corruption exceeds the limit, they are turned into a Horde Unit (the object is a variable under movement.cs). Additionally, I set it so that CameraMovement.cs, ChaseAI.cs, and HordeSpawner.cs (all 3 scripts that use the player character object) become disabled as soon as the player dies, since the destruction of the player object was causing MissingReferenceExceptions.