Rudimentary WaterMage implementation
I created a rudimentary version of the Water Mage by copy-pasting the Fire Mage object. For now, it functions identically to the fire mage. I also updated CameraMovement.cs so that the camera is placed between them. For now, everything else (HordeSpawner.cs, ChaseAi.cs) still uses the fire mage's position. Gonna hold off on adding the other 2 in case I unintentionally introduced a problem or in case there's a more optimal way to do it.